+++ django/climate_change_api/indicators/params.py
@@ -1,7 +1,7 @@
 from collections import OrderedDict
 
 from django.core.exceptions import ValidationError
-from climate_data.models import HistoricDateRange
+from climate_data.models import HistoricDateRange, ClimateDataset
 
 from .unit_converters import TemperatureConverter
 from .validators import (ChoicesValidator,
@@ -12,6 +12,9 @@ from .validators import (ChoicesValidator,
 MODELS_PARAM_DOCSTRING = ("A list of comma separated model names to filter the indicator by. The "
                           "indicator values in the response will only use the selected models. If "
                           "not provided, defaults to all models.")
+
+DATASET_PARAM_DOCSTRING = ("A single value defining which provider to use for raw climate data. "
+                           "If not provided, defaults to NEX-GDDP.")
 
 YEARS_PARAM_DOCSTRING = ("A list of comma separated year ranges to filter the response by. "
                          "Defaults to all years available. A year range is of the form "
@@ -49,7 +52,7 @@ PERCENTILE_PARAM_DOCSTRING = ("The percentile threshold used to determine the ap
 HISTORIC_RANGE_PARAM_DOCSTRING = ("The 30 year range of past years used to define the historic "
                                   "norm. Get the available options by querying the "
                                   "historic ranges endpoint. Defaults to the most recent "
-                                  "period in the historical dataset.")
+                                  "period in the historical data.")
 
 BASETEMP_PARAM_DOCSTRING = ("The base temperature used to calculate the daily difference for degree"
                             " days summations. Defaults to 65. See the 'basetemp_units' for a "
@@ -77,15 +80,36 @@ class IndicatorParam(object):
     @param default: Default value to use for parameter if none provided
     @param validators: Array of functions or classes implementing the django.core.validators
                        interface, used to validate the parameter.
+    @param empty_value: Input value that signals an empty value.
     """
 
-    def __init__(self, name, description='', required=True, default=None, validators=None):
+    def __init__(self, name, description='', required=True, default=None, validators=None,
+                 empty_value=''):
         self.name = name
         self.description = description
         self.required = required
         self.default = default
         self.validators = validators if validators is not None else []
         self.value = None
+        self.serialized_value = None
+        self.empty_value = empty_value
+
+    def set_value(self, serialized_value):
+        """Set value of parameter.
+
+        Deserialize the passed value and save both it's serialized and deserialized
+        representation after validation is passed.
+        """
+        if serialized_value is None:
+            serialized_value = self.default
+        deserialized = self.deserialize(serialized_value)
+        self.validate(deserialized)
+        self.serialized_value = serialized_value
+        self.value = deserialized
+
+    def deserialize(self, value):
+        """Deserialize a value and return the result."""
+        return value if value != self.empty_value else None
 
     def validate(self, value):
         """Validate the parameter by running all defined validators.
@@ -94,19 +118,11 @@ class IndicatorParam(object):
 
         Raises django.core.exceptions.ValidationError on the first failed validation check.
         """
-        value = value if value is not None else self.default
-
-        # for user readability, params defaulting to all options available accept 'all' as a value,
-        # which must be converted to None/null to return all
-        if value == 'all' and self.name in ('years', 'models'):
-            value = None
-
         if value is None and self.required:
             raise ValidationError('{} is required.'.format(self.name))
 
         for v in self.validators:
             v(value)
-        self.value = value
 
     def to_dict(self):
         """Return complete representation of this class as a serializable dict.
@@ -128,6 +144,12 @@ class IndicatorParam(object):
         return str(self.to_dict())
 
 
+class CommaSeparatedIndicatorParam(IndicatorParam):
+    def deserialize(self, value):
+        """Deserialize a value and return the result."""
+        return value.split(',') if value != self.empty_value else []
+
+
 class IndicatorParams(object):
     """Superclass used to define parameters necessary for an Indicator class to function.
 
@@ -135,21 +157,23 @@ class IndicatorParams(object):
     if the IndicatorParam in question has no run-time dependencies.
     """
 
-    models = IndicatorParam('models',
-                            description=MODELS_PARAM_DOCSTRING,
-                            required=False,
-                            default='all',
-                            validators=None)
-    years = IndicatorParam('years',
-                           description=YEARS_PARAM_DOCSTRING,
-                           required=False,
-                           default='all',
-                           validators=None)
-    agg = IndicatorParam('agg',
-                         description=AGG_PARAM_DOCSTRING,
-                         required=False,
-                         default='min,max,avg',
-                         validators=None)
+    models = CommaSeparatedIndicatorParam('models',
+                                          description=MODELS_PARAM_DOCSTRING,
+                                          required=False,
+                                          default='all',
+                                          validators=None,
+                                          empty_value='all')
+    years = CommaSeparatedIndicatorParam('years',
+                                         description=YEARS_PARAM_DOCSTRING,
+                                         required=False,
+                                         default='all',
+                                         validators=None,
+                                         empty_value='all')
+    agg = CommaSeparatedIndicatorParam('agg',
+                                       description=AGG_PARAM_DOCSTRING,
+                                       required=False,
+                                       default='min,max,avg',
+                                       validators=None)
 
     custom_time_agg = IndicatorParam('custom_time_agg',
                                      description=CUSTOM_TIME_AGG_PARAM_DOCSTRING,
@@ -175,12 +199,17 @@ class IndicatorParams(object):
                                                required=False,
                                                default='yearly',
                                                validators=[valid_aggregations_validator])
+        self.dataset = IndicatorParam('dataset',
+                                      description=DATASET_PARAM_DOCSTRING,
+                                      required=False,
+                                      default='NEX-GDDP',
+                                      validators=[ChoicesValidator(ClimateDataset.datasets())])
 
-    def validate(self, parameters):
-        """Validate all parameters."""
+    def set_parameters(self, parameters):
+        """Set all parameters."""
         for param_class in self._get_params_classes():
             value = parameters.get(param_class.name, None)
-            param_class.validate(value)
+            param_class.set_value(value)
 
     def to_dict(self):
         return [c.to_dict() for c in self._get_params_classes()]